refactor(queue-mobile): split into sibling components for 500-line cap
`frontend/src/pages/Queue/Mobile.tsx` was 565 lines, over the 500-line
soft cap. Decomposed into:

- mobileTypes.ts      — types, constants, formatters
- MobileRunCard.tsx   — single run card with badge
- MobileRunDetail.tsx — BottomSheet with run metadata + actions

Main `Mobile.tsx` (now 304 lines) keeps polling, cancel/rerun, and
composes the sub-components. Public export `QueueMobile` is preserved.

No behavior change.
